pyproject.toml
[project]
name = "inmindcnn"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
